+++ src/pipecatcloud/cli/entry_point.py
@@ -12,6 +12,7 @@ from loguru import logger
 from pipecatcloud._utils.console_utils import console
 from pipecatcloud.cli.commands.agent import agent_cli
 from pipecatcloud.cli.commands.auth import auth_cli
+from pipecatcloud.cli.commands.build import build_cli
 from pipecatcloud.cli.commands.deploy import create_deploy_command
 from pipecatcloud.cli.commands.docker import create_docker_command
 from pipecatcloud.cli.commands.organizations import organization_cli
@@ -78,6 +79,7 @@ def cli(
 create_deploy_command(entrypoint_cli_typer)
 create_docker_command(entrypoint_cli_typer)
 entrypoint_cli_typer.add_typer(auth_cli, rich_help_panel="Commands")
+entrypoint_cli_typer.add_typer(build_cli, rich_help_panel="Commands")
 entrypoint_cli_typer.add_typer(organization_cli, rich_help_panel="Commands")
 entrypoint_cli_typer.add_typer(regions_cli, rich_help_panel="Commands")
 entrypoint_cli_typer.add_typer(secrets_cli, rich_help_panel="Commands")